+++ src/rendering/renderer2d.cpp
@@ -124,7 +124,7 @@ namespace fea
         glViewport(viewPos.x, viewPos.y, (GLsizei)viewSize.x, (GLsizei)viewSize.y);
 
         Projection proj;
-        mProjection = proj.createOrthoProjection(0.0f, viewSize.x, 0.0f, viewSize.y, 0.000000001f, 100.0f);
+        mProjection = proj.createOrthoProjection(0.0f, (GLfloat)viewSize.x, 0.0f, (GLfloat)viewSize.y, 0.000000001f, 100.0f);
     }
     
     Viewport& Renderer2D::getViewport()
